+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added loaders for `obsFeatureMatrix.clusters.json`
 - Added loaders for `obsFeatureMatrix.genes.json`
 - Added loaders for `obsLabels.cells.json`, `obsEmbedding.cells.json`, `obsLocations.cells.json`, `obsSegmentations.cells.json`
+- Added loaders for `image.raster.json`,`obsSegmentations.raster.json`, `image.ome-zarr`
 - Added loaders for `obsLabels.anndata-expression-matrix.zarr`, `featureLabels.anndata-expression-matrix.zarr`, `obsFeatureMatrix.anndata-expression-matrix.zarr`
 
 ### Changed